vizzy_reddit/requirements.txt
praw
OpenAI
python_dotenv
requests